+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- (API) Method to check if an entity is a blight
 ### Changed
 - The "all mobs are blights" config will now work even if max difficulty is zero
 ### Fixed